// lib/types/HttpArchiveFormat.ts
export interface QueryString {
	name: string;
	value: string;
	/**  A comment provided by the user or the application */
	comment?: string | undefined;
}
// ...
export interface Param {
	/** name of a posted parameter. */
	name: string;
	/** value of a posted parameter or content of a posted file */
	value?: string | undefined;
	/** name of a posted file. */
	fileName?: string | undefined;
	/** content type of a posted file. */
	contentType?: string | undefined;
	/**  A comment provided by the user or the application */
	comment?: string | undefined;
}

/**
 * This object describes posted data, if any (embedded in `request` object).
 *
 * http://www.softwareishard.com/blog/har-12-spec/#postData
 */
export type PostData = PostDataCommon & (PostDataParams | PostDataText);

/**
 * The common properties of PostData
 */
export interface PostDataCommon {
    /** Mime type of posted data. */
    mimeType: string;
    /**  A comment provided by the user or the application */
    comment?: string | undefined;
}

/**
 * Post data with `params` specified.
 */
export interface PostDataParams {
    /**
     * List of posted parameters (in case of URL encoded parameters).
     */
    params: Param[];

    /**
     * _`params` and `text` fields are mutually exclusive._
     */
    text?: never | undefined;
}

/**
 * Post data with `text` specified.
 */
export interface PostDataText {
    /**
     * Plain text posted data
     */
    text: string;

    /**
     * _`params` and `text` fields are mutually exclusive._
     */
    params?: never | undefined;
}